requirements.txt
Flask==1.1.1
pdf2image==1.1.1
pytesseract==0.3.10
Pillow==9.1.1
